feat(exception-handling & standardized response): Integrated custom and global exception handling and standardized response structure.

+++ grocery-booking-api/src/main/java/com/qp/grocerybooking/services/impl/GroceryItemServiceImpl.java
@@ -2,10 +2,13 @@ package com.qp.grocerybooking.services.impl;
 
 import java.math.BigInteger;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.qp.grocerybooking.constants.ResponseMessages;
+import com.qp.grocerybooking.dto.response.ApiResponseDto;
 import com.qp.grocerybooking.entities.GroceryItem;
 import com.qp.grocerybooking.exceptions.ResourceNotFoundException;
 import com.qp.grocerybooking.repositories.GroceryItemRepository;
@@ -17,29 +20,38 @@ public class GroceryItemServiceImpl implements GroceryItemService{
     @Autowired
     private GroceryItemRepository groceryItemRepository;
 
-    public GroceryItem addGroceryItem(GroceryItem item) {
-        return groceryItemRepository.save(item);
+    public ApiResponseDto<GroceryItem> addGroceryItem(GroceryItem item) {
+        GroceryItem savedGroceryItem = groceryItemRepository.save(item);
+        return ApiResponseDto.<GroceryItem>builder().isSuccess(true).message(ResponseMessages.ITEM_ADDED_SUCCESSFULLY).data(savedGroceryItem).build();
     }
 
-    public List<GroceryItem> getAllGroceryItems() {
-        return groceryItemRepository.findAll();
+    public ApiResponseDto<List<GroceryItem>> getAllGroceryItems() {
+        List<GroceryItem> groceryItems = groceryItemRepository.findAll();
+        return ApiResponseDto.<List<GroceryItem>>builder().isSuccess(true).message(ResponseMessages.DATA_FETCHED_SUCCESSFULLY).data(groceryItems).build();
     }
 
-    public void deleteGroceryItem(Long id) {
+    public ApiResponseDto<String> deleteGroceryItem(Long id) {
+    	Optional<GroceryItem> groceryItem = groceryItemRepository.findById(id);
+    	if(groceryItem.isEmpty()) {
+    		throw new ResourceNotFoundException(ResponseMessages.ITEM_DOES_NOT_EXIST);
+    	}
         groceryItemRepository.deleteById(id);
+		return ApiResponseDto.<String>builder().isSuccess(true).message(ResponseMessages.ITEM_DELETED_SUCCESSFULLY).build();
     }
 
-    public GroceryItem updateGroceryItem(Long id, GroceryItem updatedItem) {
-    	GroceryItem item = groceryItemRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Item not found"));
+    public ApiResponseDto<GroceryItem> updateGroceryItem(Long id, GroceryItem updatedItem) {
+    	GroceryItem item = groceryItemRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(ResponseMessages.ITEM_DOES_NOT_EXIST));
         item.setName(updatedItem.getName());
         item.setPrice(updatedItem.getPrice());
         item.setDescription(updatedItem.getDescription());
-        return groceryItemRepository.save(item);
+        GroceryItem savedGroceryItem = groceryItemRepository.save(item);
+        return ApiResponseDto.<GroceryItem>builder().isSuccess(true).message(ResponseMessages.ITEM_UPDATED_SUCCESSFULLY).data(savedGroceryItem).build();
     }
 
-    public GroceryItem updateInventory(Long id, BigInteger quantity) {
-        GroceryItem item = groceryItemRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Item not found"));
+    public ApiResponseDto<GroceryItem> updateInventory(Long id, BigInteger quantity) {
+        GroceryItem item = groceryItemRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(ResponseMessages.ITEM_DOES_NOT_EXIST));
         item.setQuantity(quantity);
-        return groceryItemRepository.save(item);
+        GroceryItem savedGroceryItem = groceryItemRepository.save(item);
+        return ApiResponseDto.<GroceryItem>builder().isSuccess(true).message(ResponseMessages.INVENTORY_UPDATED_SUCCESSFULLY).data(savedGroceryItem).build();
     }
 }
